add flower for celery monitoring tools
