+++ django_project/core/urls.py
@@ -4,20 +4,8 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import path, include, re_path
-from django.views.generic.base import RedirectView
 
-from gap.models.preferences import Preferences
-
-
-class PreferencesRedirectView(RedirectView):
-    """Redirect to preferences admin page."""
-
-    permanent = False
-
-    def get_redirect_url(self, *args, **kwargs):
-        """Return absolute URL to redirect to."""
-        Preferences.load()
-        return '/admin/gap/preferences/1/change/'
+from core.views import PreferencesRedirectView, FlowerProxyView
 
 
 urlpatterns = [
@@ -28,6 +16,7 @@ urlpatterns = [
         r'^admin/gap/preferences/$', PreferencesRedirectView.as_view(),
         name='index'
     ),
+    FlowerProxyView.as_url(),
     path('admin/', admin.site.urls),
     path('', include('frontend.urls')),
 ]
