+++ django_project/core/views/__init__.py
@@ -1,0 +1,2 @@
+from core.views.preferences import *  # noqa
+from core.views.flower import *  # noqa
